Treat content of text/template script tags as HTML.
This introduces a new TYPE attribute context type, for attributes which specify a MIME type. This doesn't currently change the rules used to parse the attribute value (though it could); the value is just consumed as any text attribute. However, it is examined afterwards and can be used to override the element type. In this case, instead of retaining the SCRIPT element type that was set when reading the tag name, the element type is set back to NORMAL for text/template scripts. This causes the context to shift into HTML_PCDATA instead of JS after consuming the element start end tag (that is, the contents of the node will be parsed and escaped as mixed HTML content, not Javascript).