Upgrade the Closure Compiler.
Upgraded the Closure compiler to the version tagged 20190121.
This upgrade gives us access to a number of new features
for the compiler. Most notably, we want access to the function
goog.declareModuleId. This function increases the interoperability
between ES6 and Closure. Specifically, with this function we can
import ES6 modules into existing goog.modules.

TEST=manual

Ran the following scenarios:

1. An ES6 module A and another ES6 module B that imports A.
2. An ES6 module A that depends on an existing Goog module.
   Another ES6 module B that imports A.
3. A non-ES6 module A created using a goog.provide. Another
   non-ES6 module B that goog.requires A.
4. A non-ES6 module A created with a goog.module. Another
   non-ES6 module B, created with goog.module, that
   goog.requires A.
5. An ES6 module A that uses goog.declareModuleId. A non-ES6
   goog.module B that calls goog.require on A's module Id.